refactor: split utils into a package, extract models layer, centralize constants
