+++ judgearena/estimate_elo_ratings.py
@@ -19,8 +19,9 @@ from judgearena.evaluate import (
 )
 from judgearena.generate import generate_instructions
 from judgearena.log import get_logger
+from judgearena.models import make_model
 from judgearena.repro import _to_jsonable
-from judgearena.utils import cache_function_dataframe, compute_pref_summary, make_model
+from judgearena.utils import cache_function_dataframe, compute_pref_summary
 
 if TYPE_CHECKING:
     from judgearena.config import RunConfig
